requirements.txt
PyYAML==3.*
requests==2.*
